+++ NEWS.md
@@ -1,3 +1,12 @@
+apexcharter 0.4.0
+==================
+
+* Updated ApexCharts.js to 3.36.3.
+* New chart type : range area charts
+* Facets: correctly manage secondary y axis.
+
+
+
 apexcharter 0.3.1
 ==================